+++ include/LIEF/MachO/Binary.hpp
@@ -267,6 +267,15 @@ class LIEF_API Binary : public LIEF::Binary  {
   //! @param clear    If ``true`` clear the content of the section before removing
   void remove_section(const std::string& name, bool clear = false) override;
 
+  //! Remove the section from the segment with the name
+  //! given in the first parameter and with the section's name provided in the
+  //! second parameter
+  //!
+  //! @param segname     Name of the MachO::Segment
+  //! @param secname     Name of the MachO::Section to remove
+  //! @param clear       If ``true`` clear the content of the section before removing
+  void remove_section(const std::string& segname, const std::string& secname, bool clear = false);
+
   //! Remove the given LoadCommand
   bool remove(const LoadCommand& command);